moreAuth/requirements.txt
Flask==1.1.2
grpcio==1.16.0
requests_oauthlib==1.3.0
google-cloud-datastore==1.12.0
google-auth==1.14.0
requests==2.22.0
json2html==1.3.0
google_auth_oauthlib==0.4.1
google-api-python-client==1.8.3